Allow filtering. (#30)
Allow for logic to govern whether a beverage gets refilled. Can be
useful for saving execution cost and cache space when a model is
unlikely to be queried.

This can be useful when utilizing a cache store that has an eviction
policy, to prevent legacy records that happen to have been updated from
causing a more relevant record to be evicted.